pyproject.toml
[tool.poetry]
name = "swear-words"
version = "0.1.0"
description = ""
authors = ["zolizoli <zoltan.varju@gmail.com>"]
readme = "README.md"
packages = [{include = "swear_words"}]

[tool.poetry.dependencies]
python = "^3.11"
pyicu = "^2.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
